+++ src/Cedar/Proto_IPsec.c
@@ -108,7 +108,7 @@
 // test has been passed before release this source code.
 
 
-// IPsec.c
+// Proto_IPsec.c
 // IPsec module
 
 #include "CedarPch.h"
