+++ src/Cedar/Proto_OpenVPN.h
@@ -108,11 +108,11 @@
 // test has been passed before release this source code.
 
 
-// Interop_OpenVPN.h
-// Header of Interop_OpenVPN.c
+// Proto_OpenVPN.h
+// Header of Proto_OpenVPN.c
 
-#ifndef	INTEROP_OPENVPN_H
-#define	INTEROP_OPENVPN_H
+#ifndef	PROTO_OPENVPN_H
+#define	PROTO_OPENVPN_H
 
 
 //// Constants
@@ -380,6 +380,4 @@ void OpenVpnServerUdpSetDhParam(OPENVPN_SERVER_UDP *u, DH_CTX *dh);
 
 
 
-#endif	// INTEROP_OPENVPN_H
-
-
+#endif	// PROTO_OPENVPN_H
