+++ src/Cedar/Proto_SSTP.h
@@ -108,11 +108,11 @@
 // test has been passed before release this source code.
 
 
-// Interop_SSTP.h
-// Header of Interop_SSTP.c
+// Proto_SSTP.h
+// Header of Proto_SSTP.c
 
-#ifndef	INTEROP_SSTP_H
-#define	INTEROP_SSTP_H
+#ifndef	PROTO_SSTP_H
+#define	PROTO_SSTP_H
 
 //// Constants
 #define	SSTP_URI				"/sra_{BA195980-CD49-458b-9E23-C84EE0ADCD75}/"		// SSTP HTTPS URI
@@ -259,6 +259,4 @@ void SstpSendPacket(SSTP_SERVER *s, SSTP_PACKET *p);
 bool GetNoSstp();
 void SetNoSstp(bool b);
 
-#endif	// INTEROP_SSTP_H
-
-
+#endif	// PROTO_SSTP_H
